requirements.txt
torch>=1.8.1
torchvision>=0.9.1
matplotlib>=3.3.4
seaborn>=0.11.1
opencv-python>=4.5.1.48
Pillow>=8.1.0
numpy>=1.20.1
pandas>=1.2.3
PyYAML>=5.4.1
scipy>=1.6.1
